requirements.txt
dash==2.8.1
dash_bootstrap_component==1.3.1
tweepy==4.12.1
pandas==1.4.2
plotly==5.6.0